# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/product": {
      "post": {
        "operationId": "ProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "product successfully registered"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Product"
        ]
      },
      "get": {
        "operationId": "ProductController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de produtos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Listar todos os produtos",
        "tags": [
          "Product"
        ]
      }
    },
    "/product/{publicId}": {
      "get": {
        "operationId": "ProductController_findOne",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "PublicId do produto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Buscar produto por PublicId",
        "tags": [
          "Product"
        ]
      },
      "patch": {
        "operationId": "ProductController_update",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Atualiza os dados do produto",
        "tags": [
          "Product"
        ]
      },
      "delete": {
        "operationId": "ProductController_delete",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Product"
        ]
      }
    },
    "/product/{publicId}/status": {
      "patch": {
        "operationId": "ProductController_updateStatus",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Atualiza o status de um produto (ativo/inativo)",
        "tags": [
          "Product"
        ]
      }
    },
    "/product/dashboard/stock": {
      "get": {
        "operationId": "ProductController_getStockDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Product"
        ]
      }
    },
...
    "/collection": {
      "get": {
        "operationId": "CollectionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Collection"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Collection"
        ]
      },
      "post": {
        "operationId": "CollectionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "collection successfully registered"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Collection"
        ]
      }
    },
    "/collection/{publicId}": {
      "get": {
        "operationId": "CollectionController_getCollectionDetails",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "UUID da coleção",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection Details by PublicId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Collection"
        ]
      },
      "patch": {
        "operationId": "CollectionController_update",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coleção atualizada com sucesso"
          },
          "404": {
            "description": "Coleção não encontrada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Atualiza os dados da coleção",
        "tags": [
          "Collection"
        ]
      },
      "delete": {
        "operationId": "CollectionController_delete",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coleção excluida com sucesso"
          },
          "404": {
            "description": "Coleção nao encontrada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Exclui uma coleção",
        "tags": [
          "Collection"
        ]
      }
    },
    "/collection/{publicId}/status": {
      "patch": {
        "operationId": "CollectionController_updateStatus",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Atualiza o status de uma coleção (ativa/inativa)",
        "tags": [
          "Collection"
        ]
      }
    },
    "/collection/{publicId}/image": {
      "patch": {
        "operationId": "CollectionController_updateVariationImages",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "description": "PublicId da variação do produto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagem atualizada com sucesso"
          },
          "404": {
            "description": "Coleção não encontrada"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Atualiza imagem da coleção",
        "tags": [
          "Collection"
        ]
      }
    },
...
  },
  "info": {
    "title": "Minha API",
    "description": "Descrição gerada",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateProductVariationDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string",
            "example": "#d7fa92"
          },
          "name": {
            "type": "string",
            "example": "azul royal"
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "images": {
            "examples": [
              "https://example.com/masculina.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "color",
          "name",
          "price"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Produto Teste"
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "collectionPublicId": {
            "type": "string",
            "example": "4c4b0254-6534-5960-a296-f05f537f7c53"
          },
          "variations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateProductVariationDto"
            }
          }
        },
        "required": [
          "name",
          "active",
          "collectionPublicId",
          "variations"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductStatusDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "isActive"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "Collection": {
        "type": "object",
        "properties": {}
      },
      "CollectionAggregationsDto": {
        "type": "object",
        "properties": {
          "totalProducts": {
            "type": "number"
          },
          "totalStock": {
            "type": "number"
          },
          "totalValue": {
            "type": "number"
          }
        },
        "required": [
          "totalProducts",
          "totalStock",
          "totalValue"
        ]
      },
      "CollectionDetailsDto": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "aggregations": {
            "$ref": "#/components/schemas/CollectionAggregationsDto"
          }
        },
        "required": [
          "publicId",
          "name",
          "description",
          "imageUrl",
          "active",
          "createdAt",
          "updatedAt",
          "aggregations"
        ]
      },
      "CreateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Coleção Teste"
          },
          "descripttion": {
            "type": "string",
            "example": "Descrição da coleção"
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/masculina.jpg"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCollectionStatusDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "isActive"
        ]
      },
      "UpdateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UpdateCollectionImageDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string"
          }
        },
        "required": [
          "image"
        ]
      },
...
    }
  }
}